requirements.txt
Flask
Flask-CORS
gunicorn
python-dotenv
google-generativeai
waitress